hifi_code_idefics/requirements.txt
datasets==2.20.0
deepspeed==0.15.1
evaluate==0.4.3
hydra-core==1.3.2
inflection==0.5.1
nltk==3.9.1
numpy==1.26.4
omegaconf==2.3.0
peft==0.14.0
Pillow==10.4.0
pycocoevalcap==1.2
pycocotools==2.0.8
pytest==8.3.3
pytorch_lightning==2.4.0
rich>=10.2.2
setuptools>=70.0.0
termcolor==2.5.0
bytecode==0.16.1
torch==2.5.1
tqdm==4.66.5
transformers==4.45.2
wandb>=0.12.10
xxhash==3.4.1
